fix: 댓글(Comment) 엔티티 클래스의 투표 컬럼 타입 변경

+++ src/main/java/com/kakao/saramaracommunity/comment/entity/Comment.java
@@ -31,13 +31,13 @@ public class Comment extends BaseTimeEntity {
     @Column(nullable = false, columnDefinition = "TEXT")
     private String content;
 
-    private Long pick;
+    private int pick;
 
     @Column(name = "attachment_url")
     private String attachmentUrl;
 
     @Builder
-    private Comment(Member member, Board board, String content, Long pick, String attachmentUrl) {
+    private Comment(Member member, Board board, String content, int pick, String attachmentUrl) {
         this.member = member;
         this.board = board;
         this.content = content;
@@ -45,7 +45,7 @@ public class Comment extends BaseTimeEntity {
         this.attachmentUrl = attachmentUrl;
     }
 
-    public void changeComment(String content, Long pick, String attachmentUrl) {
+    public void changeComment(String content, int pick, String attachmentUrl) {
         this.content = content;
         this.pick = pick;
         this.attachmentUrl = attachmentUrl;
